# swagger/docs/swagger.yaml
definitions:
...
  dao.ResultGroupByDeclareType:
    properties:
      award_num:
        description: 获奖人数
        type: integer
      declare_num:
        description: 申报数量
        type: integer
      declare_type:
        description: 1:幼教、2:小学、3:初中、4:高中、5:职校、6:教研
        type: integer
      first_prize_num:
        description: 一等奖
        type: integer
      school_num:
        description: 申报学校数量
        type: integer
      second_prize_num:
        description: 二等奖
        type: integer
      status:
        description: 排名总数，判断是否提交状态，0：未提交，1：已提交
        type: integer
      third_prize_num:
        description: 三等奖
        type: integer
    type: object
  dao.SchoolAwardNum:
    properties:
      award_num:
        description: 获奖人数
        type: integer
      school_id:
        description: 学校id，-1：其他学校
        type: string
      school_name:
        description: 学校名称
        type: string
    type: object
  dao.TeacherTypeAwardNum:
    properties:
      award_num:
        description: 获奖人数
        type: integer
      declare_type:
        description: 1:幼教、2:小学、3:初中、4:高中、5:职校、6:教研
        type: integer
    type: object
...
  dao.YearDeclareAwardNum:
    properties:
      award_num:
        description: 获奖人数
        type: integer
      declare_num:
        description: 申报数量
        type: integer
      year:
        type: integer
    type: object
...
  resp.GetAwardRateResp:
    properties:
      award_num:
        description: 获奖人数
        type: integer
      declare_num:
        description: 申报数量
        type: integer
    type: object
...
  resp.GetDeclareRateResp:
    properties:
      declare_num:
        description: 申报数量
        type: integer
      sum_num:
        description: 总人数
        type: integer
    type: object
...
  resp.GetEverySchoolAwardNumResp:
    properties:
      list:
        items:
          $ref: '#/definitions/dao.SchoolAwardNum'
        type: array
    type: object
  resp.GetEveryTeacherTypeAwardNumResp:
    properties:
      list:
        items:
          $ref: '#/definitions/dao.TeacherTypeAwardNum'
        type: array
    type: object
  resp.GetEveryYearAwardNumResp:
    properties:
      list:
        items:
          $ref: '#/definitions/dao.YearDeclareAwardNum'
        type: array
    type: object
...
  resp.GetResultGroupByDeclareTypeResp:
    properties:
      declare_num:
        type: integer
      list:
        items:
          $ref: '#/definitions/dao.ResultGroupByDeclareType'
        type: array
      school_num:
        type: integer
    type: object
...
  resp.GetSchoolDeclareAwardRankResp:
    properties:
      list:
        items:
          $ref: '#/definitions/resp.SchoolDeclareAwardRank'
        type: array
      total:
        type: integer
    type: object
...
  resp.GetSimpleSumStatsResp:
    properties:
      list:
        items:
          $ref: '#/definitions/resp.SimpleSumStats'
        type: array
    type: object
...
  resp.GetTeacherTypeDeclareAwardRankResp:
    properties:
      list:
        items:
          $ref: '#/definitions/resp.TeacherTypeDeclareAwardRank'
        type: array
    type: object
...
  resp.GetYearDeclareAwardRankResp:
    properties:
      list:
        items:
          $ref: '#/definitions/resp.YearDeclareAwardRank'
        type: array
    type: object
...
  resp.Response:
    properties:
      code:
        type: integer
      data: {}
      error_stack: {}
      msg:
        type: string
    type: object
...
  resp.SchoolDeclareAwardRank:
    properties:
      award_num:
        description: 获奖人数
        type: integer
      award_on_year:
        description: 获奖同比
        type: number
      award_rate:
        description: 获奖占比
        type: number
      declare_num:
        description: 申报数量
        type: integer
      declare_on_year:
        description: 申报同比
        type: number
      first_prize_num:
        description: 一等奖
        type: integer
      school_id:
        description: 学校id
        type: string
      school_name:
        description: 学校名称
        type: string
      second_prize_num:
        description: 二等奖
        type: integer
      third_prize_num:
        description: 三等奖
        type: integer
    type: object
  resp.SimpleSumStats:
    properties:
      num:
        description: 数量
        type: integer
      type:
        description: 1:学校，2：老师，3：获奖，4：申报
        type: integer
      year_on_year:
        description: 同比
        type: number
    type: object
...
  resp.TeacherTypeDeclareAwardRank:
    properties:
      award_num:
        description: 获奖人数
        type: integer
      award_on_year:
        description: 获奖同比
        type: number
      award_rate:
        description: 获奖占比
        type: number
      declare_num:
        description: 申报数量
        type: integer
      declare_on_year:
        description: 申报同比
        type: number
      declare_type:
        description: 1:幼教、2:小学、3:初中、4:高中、5:职校、6:教研
        type: integer
      first_prize_num:
        description: 一等奖
        type: integer
      second_prize_num:
        description: 二等奖
        type: integer
      third_prize_num:
        description: 三等奖
        type: integer
    type: object
...
  resp.YearDeclareAwardRank:
    properties:
      award_num:
        description: 获奖人数
        type: integer
      award_on_year:
        description: 获奖同比
        type: number
      award_rate:
        description: 获奖占比
        type: number
      declare_num:
        description: 申报数量
        type: integer
      declare_on_year:
        description: 申报同比
        type: number
      first_prize_num:
        description: 一等奖
        type: integer
      num:
        description: 数量
        type: integer
      second_prize_num:
        description: 二等奖
        type: integer
      third_prize_num:
        description: 三等奖
        type: integer
      year:
        description: 年份
        type: integer
    type: object
...
host: localhost:7850
info:
  contact:
    email: support@swagger_controller.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a teachers-awards server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger teachers-awards API
  version: "1.0"
paths:
...
  /v1/statistics/activity/result:
    get:
      parameters:
      - description: 活动id
        in: query
        name: activity_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetResultGroupByDeclareTypeResp'
              type: object
      summary: 获取活动统计结果
      tags:
      - Statistics
  /v1/statistics/award/every/school:
    get:
      parameters:
      - description: 年份
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetEverySchoolAwardNumResp'
              type: object
      summary: 各校获奖情况
      tags:
      - Statistics
  /v1/statistics/award/every/year:
    get:
      parameters:
      - description: 学校id
        in: query
        name: school_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetEveryYearAwardNumResp'
              type: object
      summary: 各年获奖情况
      tags:
      - Statistics
  /v1/statistics/award/teacher/every/type:
    get:
      parameters:
      - description: 年份
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetEveryTeacherTypeAwardNumResp'
              type: object
      summary: 各类教师获奖情况
      tags:
      - Statistics
  /v1/statistics/data/detail/school:
    get:
      parameters:
      - description: 学校名称
        in: query
        name: school_name
        type: string
      - description: 年份
        in: query
        name: year
        required: true
        type: integer
      - description: 页数
        in: query
        name: page
        required: true
        type: integer
      - description: 每页大小
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetSchoolDeclareAwardRankResp'
              type: object
      summary: 各校获奖情况（具体）
      tags:
      - Statistics
  /v1/statistics/data/detail/teacher/type:
    get:
      parameters:
      - description: 年份
        in: query
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetTeacherTypeDeclareAwardRankResp'
              type: object
      summary: 各类教师获奖情况（具体）
      tags:
      - Statistics
  /v1/statistics/data/detail/year:
    get:
      parameters:
      - description: 学校id
        in: query
        name: school_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetYearDeclareAwardRankResp'
              type: object
      summary: 总表（数据统计）
      tags:
      - Statistics
  /v1/statistics/rate/award:
    get:
      parameters:
      - description: 年份
        in: query
        name: year
        type: integer
      - description: 学校id
        in: query
        name: school_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetAwardRateResp'
              type: object
      summary: 获取获奖比例
      tags:
      - Statistics
  /v1/statistics/rate/declare:
    get:
      parameters:
      - description: 年份
        in: query
        name: year
        type: integer
      - description: 学校id
        in: query
        name: school_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetDeclareRateResp'
              type: object
      summary: 获取申报比例
      tags:
      - Statistics
  /v1/statistics/simple/sum/stats:
    get:
      parameters:
      - description: 年份
        in: query
        name: year
        required: true
        type: integer
      - description: 学校id
        in: query
        name: school_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetSimpleSumStatsResp'
              type: object
      summary: 获取首页下总的统计
      tags:
      - Statistics
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"